Refactor commands into separate classes/files
