+++ flux_local/tool/manifest.py
@@ -1,0 +1,18 @@
+"""Command line tool for building, diffing, validating flux local repositories."""
+
+import pathlib
+
+from flux_local import git_repo
+
+
+class ManifestAction:
+    """Flux-local manifest action."""
+
+    async def run(  # type: ignore[no-untyped-def]
+        self,
+        path: pathlib.Path,
+        **kwargs,  # pylint: disable=unused-argument
+    ) -> None:
+        """Async Action implementation."""
+        manifest = await git_repo.build_manifest(path)
+        print(manifest.yaml())
